pkg: rename base -> common
