+++ pkg/strategy/common/strategy.go
@@ -1,4 +1,4 @@
-package base
+package common
 
 import (
 	"context"
