+++ pkg/strategy/scmaker/strategy.go
@@ -12,7 +12,7 @@ import (
 	"github.com/c9s/bbgo/pkg/fixedpoint"
 	"github.com/c9s/bbgo/pkg/indicator"
 	"github.com/c9s/bbgo/pkg/risk/riskcontrol"
-	"github.com/c9s/bbgo/pkg/strategy/base"
+	"github.com/c9s/bbgo/pkg/strategy/common"
 	"github.com/c9s/bbgo/pkg/types"
 )
 
@@ -37,7 +37,7 @@ func init() {
 
 // Strategy scmaker is a stable coin market maker
 type Strategy struct {
-	*base.Strategy
+	*common.Strategy
 
 	Environment *bbgo.Environment
 	Market      types.Market
@@ -100,7 +100,7 @@ func (s *Strategy) Subscribe(session *bbgo.ExchangeSession) {
 }
 
 func (s *Strategy) Run(ctx context.Context, orderExecutor bbgo.OrderExecutor, session *bbgo.ExchangeSession) error {
-	s.Strategy = &base.Strategy{}
+	s.Strategy = &common.Strategy{}
 	s.Strategy.Setup(ctx, s.Environment, session, s.Market, ID, s.InstanceID())
 
 	s.book = types.NewStreamBook(s.Symbol)
